+++ packages/eslint-plugin/configs/custom.js
@@ -4,8 +4,9 @@ module.exports = {
 		'@wordpress/no-unused-vars-before-return': 'error',
 		'@wordpress/no-base-control-with-label-without-id': 'error',
 		'@wordpress/no-unguarded-get-range-at': 'error',
-		'@wordpress/no-global-active-element': 'warn',
+		'@wordpress/no-global-active-element': 'error',
 		'@wordpress/no-global-get-selection': 'error',
+		'@wordpress/no-global-event-listener': 'warn',
 	},
 	overrides: [
 		{
@@ -23,6 +24,7 @@ module.exports = {
 			rules: {
 				'@wordpress/no-global-active-element': 'off',
 				'@wordpress/no-global-get-selection': 'off',
+				'@wordpress/no-global-event-listener': 'off',
 			},
 		},
 	],