+++ packages/react-native-editor/CHANGELOG.md
@@ -10,6 +10,12 @@ For each user feature we should also add a importance categorization label  to i
 -->
 
 ## Unreleased
+
+* [***] Adding support for selecting different unit of value in Cover and Columns blocks
+* [**] Button block - Add link picker to the block settings
+
+## 1.41.0
+
 * [***] Faster editor start and overall operation on Android [#26732]
 
 * [*] [Android] Enable multiple upload support for Image block